+++ app/config-types.ts
@@ -795,10 +795,45 @@ const MapSymbolLayer = t.type({
 });
 export type MapSymbolLayer = t.TypeOf<typeof MapSymbolLayer>;
 
+const BaseCustomLayer = t.type({
+  id: t.string,
+  isBehindAreaLayer: t.boolean,
+  syncTemporalFilters: t.boolean,
+});
+export type BaseCustomLayer = t.TypeOf<typeof BaseCustomLayer>;
+
+const WMSCustomLayer = t.intersection([
+  t.type({
+    type: t.literal("wms"),
+  }),
+  BaseCustomLayer,
+]);
+export type WMSCustomLayer = t.TypeOf<typeof WMSCustomLayer>;
+export const getWMSCustomLayers = (
+  customLayers: BaseLayer["customLayers"]
+): WMSCustomLayer[] => {
+  return customLayers.filter((l) => l.type === "wms") as WMSCustomLayer[];
+};
+
+const WMTSCustomLayer = t.intersection([
+  t.type({
+    type: t.literal("wmts"),
+    url: t.string,
+  }),
+  BaseCustomLayer,
+]);
+export type WMTSCustomLayer = t.TypeOf<typeof WMTSCustomLayer>;
+export const getWMTSCustomLayers = (
+  customLayers: BaseLayer["customLayers"]
+): WMTSCustomLayer[] => {
+  return customLayers.filter((l) => l.type === "wmts") as WMTSCustomLayer[];
+};
+
 const BaseLayer = t.type({
   show: t.boolean,
   locked: t.boolean,
   bbox: t.union([BBox, t.undefined]),
+  customLayers: t.array(t.union([WMSCustomLayer, WMTSCustomLayer])),
 });
 export type BaseLayer = t.TypeOf<typeof BaseLayer>;
 
@@ -1071,7 +1106,8 @@ export const isColorInConfig = (
   | ColumnConfig
   | LineConfig
   | PieConfig
-  | ScatterPlotConfig => {
+  | ScatterPlotConfig
+  | BarConfig => {
   return !isTableConfig(chartConfig) && !isMapConfig(chartConfig);
 };