+++ app/configurator/components/chart-controls/color-picker.tsx
@@ -46,7 +46,7 @@ export const Swatch = ({
   return (
     <Box
       className={classes.swatch}
-      data-testId="color-picker-swatch"
+      data-testid="color-picker-swatch"
       sx={{
         borderColor: selected ? borderColor : undefined,
         boxShadow: selected ? `0 0 0.5rem 0 ${color}` : undefined,