+++ app/utils/chart-config/versioning.ts
@@ -1478,6 +1478,33 @@ export const chartConfigMigrations: Migration[] = [
       return newConfig;
     },
   },
+  {
+    from: "4.2.0",
+    to: "4.3.0",
+    description: `maps {
+      baseLayer {
+        + customLayers
+      }
+    }`,
+    up: (config) => {
+      const newConfig = { ...config, version: "4.3.0" };
+
+      if (newConfig.chartType === "map") {
+        newConfig.baseLayer.customLayers = [];
+      }
+
+      return newConfig;
+    },
+    down: (config) => {
+      const newConfig = { ...config, version: "4.2.0" };
+
+      if (newConfig.chartType === "map") {
+        delete newConfig.baseLayer.customLayers;
+      }
+
+      return newConfig;
+    },
+  },
 ];
 
 export const migrateChartConfig = makeMigrate<ChartConfig>(
@@ -1544,7 +1571,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "3.0.0",
     up: async (config) => {
       const newConfig = { ...config, version: "3.0.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1562,7 +1589,7 @@ export const configuratorStateMigrations: Migration[] = [
     down: async (config) => {
       const newConfig = { ...config, version: "2.0.0" };
       let dataSet: string | undefined;
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         if (!dataSet) {
@@ -1594,7 +1621,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "3.0.1",
     up: async (config) => {
       const newConfig = { ...config, version: "3.0.1" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1610,7 +1637,7 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "3.0.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1667,7 +1694,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "3.2.0",
     up: async (config) => {
       const newConfig = { ...config, version: "3.2.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1683,7 +1710,7 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "3.1.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1750,7 +1777,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "3.4.0",
     up: async (config) => {
       const newConfig = { ...config, version: "3.4.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1766,7 +1793,7 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "3.3.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1896,7 +1923,7 @@ export const configuratorStateMigrations: Migration[] = [
         },
       };
 
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1923,7 +1950,7 @@ export const configuratorStateMigrations: Migration[] = [
         },
       };
 
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -1987,7 +2014,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "4.0.0",
     up: async (config) => {
       const newConfig = { ...config, version: "4.0.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -2027,7 +2054,7 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "3.8.0" };
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -2068,8 +2095,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "4.1.0",
     up: async (config) => {
       const newConfig = { ...config, version: "4.1.0" };
-
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -2085,8 +2111,7 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "4.0.0" };
-
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -2155,8 +2180,7 @@ export const configuratorStateMigrations: Migration[] = [
     to: "4.3.0",
     up: async (config) => {
       const newConfig = { ...config, version: "4.3.0" };
-
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
@@ -2172,13 +2196,49 @@ export const configuratorStateMigrations: Migration[] = [
     },
     down: async (config) => {
       const newConfig = { ...config, version: "4.2.0" };
-
-      const chartConfigs: any[] = [];
+      const chartConfigs = [];
 
       for (const chartConfig of newConfig.chartConfigs) {
         const migratedChartConfig = await migrateChartConfig(chartConfig, {
           migrationProps: newConfig,
           toVersion: "4.1.0",
+        });
+        chartConfigs.push(migratedChartConfig);
+      }
+
+      newConfig.chartConfigs = chartConfigs;
+
+      return newConfig;
+    },
+  },
+  {
+    description: "ALL (bump ChartConfig version)",
+    from: "4.3.0",
+    to: "4.4.0",
+    up: async (config) => {
+      const newConfig = { ...config, version: "4.4.0" };
+      const chartConfigs = [];
+
+      for (const chartConfig of newConfig.chartConfigs) {
+        const migratedChartConfig = await migrateChartConfig(chartConfig, {
+          migrationProps: newConfig,
+          toVersion: "4.3.0",
+        });
+        chartConfigs.push(migratedChartConfig);
+      }
+
+      newConfig.chartConfigs = chartConfigs;
+
+      return newConfig;
+    },
+    down: async (config) => {
+      const newConfig = { ...config, version: "4.3.0" };
+      const chartConfigs = [];
+
+      for (const chartConfig of newConfig.chartConfigs) {
+        const migratedChartConfig = await migrateChartConfig(chartConfig, {
+          migrationProps: newConfig,
+          toVersion: "4.2.0",
         });
         chartConfigs.push(migratedChartConfig);
       }