+++ CHANGELOG.md
@@ -12,7 +12,24 @@ You can also check the
 # Unreleased
 
 - Features
+  - It's now possible to add custom WMS layers coming from wms.geo.admin.ch to
+    map charts
+- Fixes
+  - Limits that are connected to values that are filtered out do not render in
+    the chart anymore
+
+# [5.3.0] - 2025-02-25
+
+- Features
+  - It's now possible to add custom WMTS layers coming from wmts.geo.admin.ch to
+    map charts
   - Limits can now be displayed in bar charts
+  - Exact limit values are now displayed in legend
+- Fixes
+  - Line chart dots are now based on data, eliminating misaligned dots
+  - Added a fallback to color selector
+  - Added dynamic descriptions for custom color palette types
+  - Show line dot sizes are now translated
 
 # [5.2.6] - 2025-02-18
 
@@ -31,6 +48,7 @@ You can also check the
   - Added option for hiding legend titles using a toggle switch
   - Bar charts, dashboard text blocks and Markdown inputs are not hidden behind
     flags anymore
+  - Added improved Iframe view for the use of ODS
 - Fixes
   - Added button translations for custom color palette update and create button
   - Color swatches inside the color picker dynamically adjust to the colors of
@@ -47,6 +65,9 @@ You can also check the
   - Fixed errors regarding switching form existing categorical palette to a
     diverging color palette
   - Improved filter section styling
+  - Removed legend titles tooltip on the toggle switch
+  - Fixed Map Symbol Layer custom color palette support for all palette types
+  - Added color palettes to bar chart type
 - Maintenance
   - Added authentication method to e2e tests
   - Added e2e tests for custom color palettes